Add bibcode entry command
This adds aiohttp as a dependency. I'm adopting aiohttp as the HTTP
client library since it'll future-proof us for when we need to make
multiple asynchronous requests.

The bibcode entry command is the MVP to make bibcode worth it by using
the export API to convert a bibcode into a bib entry of a specified
format.

https://github.com/adsabs/adsabs-dev-api/blob/master/Export_API.ipynb